pipeline/requirements.txt
anthropic>=0.30.0
feedparser>=6.0.11
httpx>=0.27.0
playwright>=1.44.0
playwright-stealth>=1.0.6
googlenewsdecoder>=0.1.5
Pillow>=10.3.0
python-dotenv>=1.0.1
